The CalciteArray class

The CalciteArray is a DelayedArray wrapper around a on-disk array downloaded from the calcite backend. Specifically, it remembers the zircon ID that was used to download the array, which can be used to avoid a redundant write and upload in saveObject and uploadDirectory, respectively. The actual heavy-lifting is done by forwarding all operations to the internal DelayedArray-compatible seed object, which may be any of the usual classes, e.g., a HDF5ArraySeed, H5SparseMatrixSeed or any of its delayed operations.

CalciteArraySeed(id)

CalciteArray(id, ...)

# S4 method for CalciteArraySeed
DelayedArray(seed)

Arguments

id

String containing the calcite identifier, or a list containing the unpacked components of an identifier (from unpackID). For CalciteArray, this can also be a CalciteArraySeed object.

...

Further arguments to pass to the CalciteArraySeed constructor.

seed

A CalciteArraySeed object.

Value

The CalciteArraySeed constructor returns a CalciteArraySeed object.

The CalciteArray constructor and the DelayedArray method return a CalciteArray object (or a CalciteMatrix, for two dimensions).

Details

If id uses the latest version alias, this is automatically resolved to a full version by the CalciteArraySeed constructor. This ensures that the version is explicitly pinned in downstream applications and when saving to file.

Examples

fun <- calcite:::.configure_cache()
id <- "test:my_first_sce/assay-1/matrix.h5@v1"
mat <- CalciteArray(id)
mat
#> <20006 x 3005> sparse matrix of class CalciteMatrix and type "integer":
#>          1772071015_C02 1772071017_G12 ... 1772066098_A12 1772058148_F03
#>  Tspan12              0              0   .              0              1
#>    Tshz1              3              1   .              0              1
#>   Fnbp1l              3              1   .              0              0
#> Adamts15              0              0   .              0              0
#>   Cldn12              1              1   .              0              0
#>      ...              .              .   .              .              .
#>   mt-Co2             65             97   .             29            120
#>   mt-Co1             75            111   .             20            143
#>  mt-Rnr2            158            326   .             36            359
#>  mt-Rnr1             31             88   .             12             52
#>  mt-Nd4l             13             14   .              3             13
fun()