Add calcite annotation

Add calcite's required annotation to a Bioconductor Annotated object.

annotateObject(x, title, description, maintainers, species, genome, origin)

objectAnnotation(x)

setAnnotation(x, annotation)

Arguments

x

An Annotated object.

title

String containing the object's title.

description

String containing the description of the object.

maintainers

List containing the identity of the maintainers. Each entry may be a person-formatted string, or a list containing the name and email strings.

Alternatively, a character vector of person-formatted strings.

species

An integer vector of NCBI taxonomy IDs for the species relevant to the data in x.

genome

List describing the genomes involved in constructing x. Each entry should be a list containing:

id, a string containing the genome build ID (e.g., hg38, NCBIm37)

type, a string specifying the type of ID (e.g., UCSC, Ensembl)

Check out https://artifactdb.github.io/calcite-schemas/array/v1.html#allOf_i0_genome for details.

origin

List describing the origin of x. Each entry should be a list containing:

source, a string describing the source repository, e.g., PubMed, GEO, ArrayExpress.

id, a string containing an identifier within source.

Check out https://artifactdb.github.io/calcite-schemas/array/v1.html#allOf_i0_origin for details.

annotation

List of calcite-relevant metadata.

Value

For annotateObject, x is returned with extra fields in its metadata.

For objectAnnotation, a list of calcite-relevant metadata is returned.

For setannotation, x is returned after replacing the calcite-relevant metadata with annotation.

Author

Aaron Lun

Examples

library(S4Vectors)
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
df <- DataFrame(X=1:10, Y=LETTERS[1:10], Z=factor(letters[1:10]))

df <- annotateObject(df,
    title="FOO",
    description="I am a data frame",
    maintainers="Aaron Lun <infinite.monkeys.with.keyboards@gmail.com>",
    species=9606,
    genome=list(list(id="hg38", source="UCSC")),
    origin=list(list(source="PubMed", id="123456789"))
)

anno <- objectAnnotation(df)
str(anno)
#> List of 7
#>  $ title       : chr "FOO"
#>  $ description : chr "I am a data frame"
#>  $ maintainers :List of 1
#>   ..$ :List of 2
#>   .. ..$ name : chr "Aaron Lun"
#>   .. ..$ email: chr "infinite.monkeys.with.keyboards@gmail.com"
#>  $ species     : num 9606
#>  $ genome      :List of 1
#>   ..$ :List of 2
#>   .. ..$ id    : chr "hg38"
#>   .. ..$ source: chr "UCSC"
#>  $ origin      :List of 1
#>   ..$ :List of 2
#>   .. ..$ source: chr "PubMed"
#>   .. ..$ id    : chr "123456789"
#>  $ bioc_version: chr "3.16"

anno$maintainers <- c(anno$maintainers, 
    list(list(name="Darth Vader", email="vader@empire.gov")))
df <- setAnnotation(df, anno)