Create a InputTimePicker component

Create a InputTimePicker component

Usage

calcite_input_time_picker(...)

Arguments

...

named attributes passed to htmltools::tag()

Value

an object of class calcite_component which is a subclass of shiny.tag

Details

Properties

The following properties are provided by this component:

Name	Attribute	Description	Values	Reflects to Attribute

disabled	disabled	When true, interaction is prevented and the component is displayed with lower opacity.	boolean	TRUE

focusTrapDisabled	focus-trap-disabled	When true, prevents focus trapping.	boolean	TRUE

form	form	The id of the form that will be associated with the component. When not set, the component will be associated with its ancestor form element, if any.	string	TRUE

hourFormat	hour-format	Specifies the component's hour format, where: "user" displays the user's locale format, "12" displays a 12-hour format, and "24" displays a 24-hour format.	"12" | "24" | "user"	TRUE

max	max	When the component resides in a form, specifies the maximum value.	string	TRUE

messageOverrides	NA	Use this property to override individual strings used by the component.	Check API reference	FALSE

min	min	When the component resides in a form, specifies the minimum value.	string	TRUE

name	name	Specifies the name of the component on form submission.	string	FALSE

numberingSystem	numbering-system	Specifies the Unicode numeral system used by the component for localization.	"arab" | "arabext" | "latn"	FALSE

open	open	When true, displays the calcite-time-picker component.	boolean	TRUE

overlayPositioning	overlay-positioning	Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container's layout. "fixed" should be used to escape an overflowing parent container, or when the reference element's position CSS property is "fixed".	"absolute" | "fixed"	FALSE

placement	placement	Determines where the popover will be positioned relative to the input.	"auto" | "auto-end" | "auto-start" | "bottom" | "bottom-end" | "bottom-start" | "leading" | "leading-end" | "leading-start" | "left" | "left-end" | "left-start" | "right" | "right-end" | "right-start" | "top" | "top-end" | "top-start" | "trailing" | "trailing-end" | "trailing-start"	TRUE

readOnly	read-only	When true, the component's value can be read, but controls are not accessible and the value cannot be modified.	boolean	TRUE

required	required	When true and the component resides in a form, the component must have a value in order for the form to submit.	boolean	TRUE

scale	scale	Specifies the size of the component.	"l" | "m" | "s"	TRUE

status	status	Specifies the status of the input field, which determines message and icons.	"idle" | "invalid" | "valid"	TRUE

step	step	Specifies the granularity the component's value must adhere to (in seconds).	number	FALSE

validationIcon	validation-icon	Specifies the validation icon to display under the component.	boolean | string	TRUE

validationMessage	validation-message	Specifies the validation message to display under the component.	string	FALSE

validity	NA	The current validation state of the component.	Check API reference	FALSE

value	value	The time value in ISO (24-hour) format.	string	FALSE

Events

The following events are observed by shiny:

Event	Description

calciteInputTimePickerBeforeClose	Fires when the component is requested to be closed and before the closing transition begins.

calciteInputTimePickerBeforeOpen	Fires when the component is added to the DOM but not rendered, and before the opening transition begins.

calciteInputTimePickerChange	Fires when the component's value is changes.

calciteInputTimePickerClose	Fires when the component is closed and animation is complete.

calciteInputTimePickerOpen	Fires when the component is open and animation is complete.

References

Official Documentation

Examples

calcite_input_time_picker()
#> <calcite-input-time-picker></calcite-input-time-picker>