Fetch an object from the calcite backend

Fetch an object from the calcite backend. This involves caching its associated files before loading it into the current R session.

fetchObject(id, cache = TRUE, force.update = FALSE)

Arguments

id

String containing the ArtifactDB identifier for the object of interest.

cache

Logical scalar indicating whether caching should be performed. If TRUE, a default cache location is chosen.

Alternatively, a string specifying the path to a cache directory.

Alternatively, a BiocFileCache object.

force.update

Logical scalar indicating whether cache entries should be forcibly updated. Useful for fixing corrupted or incomplete files in the cache.

Value

An R object corresponding to id.

See also

Author

Aaron Lun

Examples

cache.dir <- tempfile()
id <- exampleID()

# First request downloads the resources:
obj <- fetchObject(id, cache=cache.dir)
obj
#> DataFrame with 10 rows and 4 columns
#>            X           Y        Z                  AA
#>    <integer> <character> <factor>         <DataFrame>
#> 1          1           A        a 0.221781: 0.1249998
#> 2          2           B        b 0.331854:-0.9385503
#> 3          3           C        c 0.807343: 1.3094641
#> 4          4           D        d 0.483355:-0.6231195
#> 5          5           E        e 0.108293:-0.0249276
#> 6          6           F        f 0.466883: 0.8491610
#> 7          7           G        g 0.243697:-0.4783389
#> 8          8           H        h 0.848236: 0.4384521
#> 9          9           I        i 0.316582: 0.2180794
#> 10        10           J        j 0.582001: 0.7062769

# Next query just uses the cache:
fetchObject(id, cache=cache.dir)
#> DataFrame with 10 rows and 4 columns
#>            X           Y        Z                  AA
#>    <integer> <character> <factor>         <DataFrame>
#> 1          1           A        a 0.221781: 0.1249998
#> 2          2           B        b 0.331854:-0.9385503
#> 3          3           C        c 0.807343: 1.3094641
#> 4          4           D        d 0.483355:-0.6231195
#> 5          5           E        e 0.108293:-0.0249276
#> 6          6           F        f 0.466883: 0.8491610
#> 7          7           G        g 0.243697:-0.4783389
#> 8          8           H        h 0.848236: 0.4384521
#> 9          9           I        i 0.316582: 0.2180794
#> 10        10           J        j 0.582001: 0.7062769

# All the annotation is attached:
str(objectAnnotation(obj))
#> List of 8
#>  $ title       : chr "FOO"
#>  $ description : chr "I am a data frame"
#>  $ maintainers :List of 1
#>   ..$ :List of 2
#>   .. ..$ name : chr "Aaron Lun"
#>   .. ..$ email: chr "infinite.monkeys.with.keyboards@gmail.com"
#>  $ species     : int 9606
#>  $ genome      :List of 1
#>   ..$ :List of 2
#>   .. ..$ id    : chr "hg38"
#>   .. ..$ source: chr "UCSC"
#>  $ origin      :List of 1
#>   ..$ :List of 2
#>   .. ..$ source: chr "PubMed"
#>   .. ..$ id    : chr "123456789"
#>  $ bioc_version: chr "3.16"
#>  $ _extra      :List of 9
#>   ..$ $schema      : chr "csv_data_frame/v1.json"
#>   ..$ id           : chr "test:my_first_df/simple.csv.gz@v1"
#>   ..$ project_id   : chr "test"
#>   ..$ version      : chr "v1"
#>   ..$ metapath     : chr "my_first_df/simple.csv.gz"
#>   ..$ meta_indexed : chr "2022-10-07T20:26:32.770Z"
#>   ..$ meta_uploaded: chr "2022-10-07T20:26:09.376Z"
#>   ..$ uploaded     : chr "2022-10-07T20:26:09.376Z"
#>   ..$ permissions  :List of 5
#>   .. ..$ scope       : chr "project"
#>   .. ..$ read_access : chr "public"
#>   .. ..$ write_access: chr "owners"
#>   .. ..$ owners      : chr "ArtifactDB-bot"
#>   .. ..$ viewers     : list()