Save an object to file

Save an object to a staging directory in preparation for upload. Multiple objects may be saved inside the same staging directory, though users should avoid saving an object inside another object's subdirectory (created by a previous saveObject call).

saveObject(x, dir, path)

Arguments

x

A supported Bioconductor object, with annotation added by annotateObject.

dir

String containing a path to a staging directory.

path

String containing the relative path inside the staging directory in which to save the object's contents. This should not be nested inside any subdirectories created by previous saveObject calls.

Value

x is saved to the specified location, and NULL is invisibly returned.

See also

Author

Aaron Lun

Examples

# Making an example DataFrame:
df <- exampleObject()
df
#> DataFrame with 10 rows and 4 columns
#>            X           Y        Z                   AA
#>    <integer> <character> <factor>          <DataFrame>
#> 1          1           A        a 0.8243776: 1.1507923
#> 2          2           B        b 0.3176201:-0.3031649
#> 3          3           C        c 0.2473154:-0.5254622
#> 4          4           D        d 0.3724730:-0.4377465
#> 5          5           E        e 0.5166181:-2.0102011
#> 6          6           F        f 0.2713153:-0.0962849
#> 7          7           G        g 0.0703847: 0.1819448
#> 8          8           H        h 0.2208868:-0.4250750
#> 9          9           I        i 0.2685402: 1.1621757
#> 10        10           J        j 0.6679586:-1.7762442
str(objectAnnotation(df))
#> List of 7
#>  $ title       : chr "FOO"
#>  $ description : chr "I am a data frame"
#>  $ maintainers :List of 1
#>   ..$ :List of 2
#>   .. ..$ name : chr "Aaron Lun"
#>   .. ..$ email: chr "infinite.monkeys.with.keyboards@gmail.com"
#>  $ species     : num 9606
#>  $ genome      :List of 1
#>   ..$ :List of 2
#>   .. ..$ id    : chr "hg38"
#>   .. ..$ source: chr "UCSC"
#>  $ origin      :List of 1
#>   ..$ :List of 2
#>   .. ..$ source: chr "PubMed"
#>   .. ..$ id    : chr "123456789"
#>  $ bioc_version: chr "3.16"

# Saving it to a directory:
tmp <- tempfile()
dir.create(tmp)
saveObject(df, tmp, "my_first_df")

list.files(tmp, recursive=TRUE)
#> [1] "my_first_df.json"                      
#> [2] "my_first_df/column3/levels.csv.gz"     
#> [3] "my_first_df/column3/levels.csv.gz.json"
#> [4] "my_first_df/column4/simple.csv.gz"     
#> [5] "my_first_df/column4/simple.csv.gz.json"
#> [6] "my_first_df/simple.csv.gz"             
#> [7] "my_first_df/simple.csv.gz.json"