Create a Panel component

Create a Panel component

Usage

calcite_panel(...)

Arguments

...

named attributes passed to htmltools::tag()

Value

an object of class calcite_component which is a subclass of shiny.tag

Details

Properties

The following properties are provided by this component:

Name	Attribute	Description	Values	Reflects to Attribute

beforeClose	NA	Passes a function to run before the component closes.	Check API reference	FALSE

closable	closable	When true, displays a close button in the trailing side of the header.	boolean	TRUE

closed	closed	When true, the component will be hidden.	boolean	TRUE

collapsed	collapsed	When true, hides the component's content area.	boolean	TRUE

collapsible	collapsible	When true, the component is collapsible.	boolean	TRUE

description	description	A description for the component.	string	FALSE

disabled	disabled	When true, interaction is prevented and the component is displayed with lower opacity.	boolean	TRUE

heading	heading	The component header text.	string	FALSE

headingLevel	heading-level	Specifies the heading level of the component's heading for proper document structure, without affecting visual styling.	1 | 2 | 3 | 4 | 5 | 6	TRUE

loading	loading	When true, a busy indicator is displayed.	boolean	TRUE

menuFlipPlacements	NA	Specifies the component's fallback menu placement when it's initial or specified placement has insufficient space available.	Check API reference	FALSE

menuOpen	menu-open	When true, the action menu items in the header-menu-actions slot are open.	boolean	TRUE

menuPlacement	menu-placement	Determines where the action menu will be positioned.	"auto" | "auto-end" | "auto-start" | "bottom" | "bottom-end" | "bottom-start" | "leading" | "leading-end" | "leading-start" | "left" | "left-end" | "left-start" | "right" | "right-end" | "right-start" | "top" | "top-end" | "top-start" | "trailing" | "trailing-end" | "trailing-start"	TRUE

messageOverrides	NA	Use this property to override individual strings used by the component.	Check API reference	FALSE

overlayPositioning	overlay-positioning	Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container's layout. "fixed" should be used to escape an overflowing parent container, or when the reference element's position CSS property is "fixed".	"absolute" | "fixed"	TRUE

scale	scale	Specifies the size of the component.	"l" | "m" | "s"	TRUE

Events

The following events are observed by shiny:

Event	Description

calcitePanelClose	Fires when the close button is clicked.

calcitePanelScroll	Fires when the content is scrolled.

calcitePanelToggle	Fires when the collapse button is clicked.

Slots

The following slots are provided by this component:

Slot	Description

Default (unnamed)	A slot for adding custom content.

action-bar	A slot for adding a calcite-action-bar to the component.

alerts	A slot for adding calcite-alerts to the component.

content-bottom	A slot for adding content below the unnamed (default) slot and above the footer slot (if populated)

content-top	A slot for adding content above the unnamed (default) slot and below the action-bar slot (if populated).

header-actions-start	A slot for adding actions or content to the start side of the header.

header-actions-end	A slot for adding actions or content to the end side of the header.

header-content	A slot for adding custom content to the header.

header-menu-actions	A slot for adding an overflow menu with actions inside a calcite-dropdown.

fab	A slot for adding a calcite-fab (floating action button) to perform an action.

footer	A slot for adding custom content to the component's footer. Should not be used with the "footer-start" or "footer-end" slots.

footer-actions	Deprecated Use the footer-start and footer-end slots instead. A slot for adding calcite-buttons to the component's footer.

footer-end	A slot for adding a trailing footer custom content. Should not be used with the "footer" slot.

footer-start	A slot for adding a leading footer custom content. Should not be used with the "footer" slot.

References

Official Documentation

Examples

calcite_panel()
#> <calcite-panel></calcite-panel>